Refactoring
pull out GameStrings

add debug output for crystal headbutt trees
fix qrpkm item access
